+++ kong/plugins/acme/reserved_words.lua
@@ -1,0 +1,7 @@
+local reserved_words = {
+  RENEW_KEY_PREFIX = "kong_acme:renew_config:",
+  RENEW_LAST_RUN_KEY = "kong_acme:renew_last_run",
+  CERTKEY_KEY_PREFIX = "kong_acme:cert_key:",
+}
+
+return reserved_words